Fix encoding issues inside the input field
The current approach casts the received request object from a byte
array to a string. Upon copying it into the UI input field, it expects
again a bytes array. This back and forth conversion will try to encode
binary blobs leading to malformed requests. As all functions expect
to work with byte arrays the cast to a string can be removed to maintain
the original requests structure