Support raw input in FormatedTextFields
FormatedTextFields did not support raw input so far. For all different
options, the input was concerted into a string first. This is now only
done for the UTF8 and Latin formats. The rest operates on raw bytes.

This was implemented because one of the cstc variables (e.g.
§Outgoing_step1) could contain a raw byte value. These were translated
to strings inside of the substitution process.