Allow bytes in Replace operator
The Replace operator from the String categroy is kind of difficult. In
order to allow regex matches we need to perform String transformations.
Therefore, I added an additional checkbox which allows switching between
regex and non regex mode. In non regex mode, bytes are allowed and are
not converted.

Late we should improve this operation. Currently, the non regex replace
only replaces the first occurence. There should be options to modify
this behavior.