Enable context menu in CSTC pane
The context menu (right click) was not working in the cstc pane. Having
this feature is nice e.g. for copy/paste and sending the result of a CSTC
transformation to the repeater for testing.

By introducing the new CstcMessageEditorController class and adding this
as the message controller for the IMessageEditorTabs used insdie the
CTSC pane, the context menu is now enabled and should work.

However, it was not fully tested that each context menu action leads to
the desired result. Furthermore, enabling the context menu in CSTC tabs
seems not to be possible yet (missing Burp API functions)