Prevent null pointer exception with empty response
If a request with no valid server response yet is send to CSTC incoming,
a null ponter exception was thrown. This was the case, because Burp
responds with ``null`` when requesting the bytes from a non present
response.

This case is now catched and ``null`` content of the response is
replaced with a generic placeholder.