Add SetIfEmpty operation
Added a SetIfEmpty operation. This operation simply takes the last
result of the current lane, checks if it is empty, and replaces it with
a constant value if this is the case.

This is useful in cases where you want a default value for a variable if
it is empty. Just place this operation at the end of the lane and set
the desired default value.