Change variable prefix to $
When installing CSTC from BApp store, it turned out that the § variable
prefix did not work anymore. It was only possible to insert variables
using the context menu, which used the two-byte representation of §.

After contacting Burp support, they said the easiest solution is to
switch to a ascii prefix. So we decided to switch the variable prefix
to $.