Allow more flexible variable replacement for bytes
When variables are used inside of byte operations, CSTC only allowed a 1
to 1 match for replacements. E.g. ``$Outgoing_step1`` was replaced, ``t:
$Outgoing_step1`` was not replaced.

This was now chaned. Also in byte oeprations variables should now be
replaced with surrounding text. It is also possible to have multiple
variables inside the same replacement.