Fix nullpointer exception on save operations
When CSTC saved a receipt, it used the caption of each element to store
current values. Checkboxes and Buttons usually have no captions
assigned, which cuased nullpointer execeptions for all receipts
containing these elements. An method to fix this was now implemented by
allowing each GUI component to have an additional identifier. This needs
now to be applied to the different operations.