Allow non-autoupdated fields
When cstc loads an reciept it tries to restore the original state. For
the keystore operation, this does not make sense. We do not store the
password for the keystore, and therefore it cannot be reopened
automatically.

By introducing a new identifier prefix 'noupdae', GUI elements can now
specify if they want to be auto-initialized or leaved blank.