Fix tooltipps for DefaultTreeNodes
CSTC implements tooltipps for its different operations. For operation
categories, the tooltipp was left empty. This lead to an empty box being
displayed when hovering over a category.

To avoid this, tooltipps for operation categories have beein disabled.
In future we will probably define useful descriptions for each category
and re-enable them. But for now they were no longer displayed.